Story/aad 44 (#11)
* Creating Info page for each shelter

Creating the get more info page for each of the shelters that are listed on the results page. Further styling is needed, and a button to direct to the mapview is needed.

* Update results.blade.php

Slugging the urls on the result page to make url more readable

* Adding images for shelters

These are temporary images for each of the shelters. Will need to talk to sponsors to get better sourced images

* Fixing image loading

Wrote a new function for getting images loaded using a variable and with disregard for file extension. These images will need to live in a different location.

* Update WelcomeController.php

Fixed images that had parentheses in their name

* Update WelcomeController.php

Fixed apostrophes

* Update WelcomeController.php

Removed all non alphanumeric chars due to special opening and closing quote characters.

* Built out the Get More Info Blade

Created and began basic styling of the get more info page that is can be navigated to from the results page.

Co-authored-by: Bishop <59637502+Iwasson@users.noreply.github.com>