Develop (#17)
* Story/aad 23 (#2)

* Initial views created

The current styling and layout are not for final product use.

* Getting sheets to work

Getting the base functionality of google sheets to work

* Work done on filters and styling

* Refactored filters

* Transitioning to Bootstrap 5.2.0

* Loading bootstrap locally

A local version of bootstrap has been loaded into the project. This will mean that we no longer need to rely on the CDN and we have a stable version of bootstrap

* Filter modal working

Automated filter population should now work, color and additional styling still needs to be done.

* Moved filter modal to its own file

Co-authored-by: Bishop <59637502+Iwasson@users.noreply.github.com>

* Story/aad 33 (#3)

* Expanding sheets controller

Added function to retrieve all shelter info from the info tab and last input from the forms.

* Creating Card List View

Created necessary routing and controller. Controller parses data from google sheets into an array of key values  that are then used to display cards contain shelter info on the frontend.

* Added working filter functionality (#5)

The submission of the search form will now display the shelters that match the criteria

* Update SheetsController.php (#9)

Offloading some of the work to the DB. This allows us to reduce the load time by reducing how many networking requests are issued. This backend change is used to reference the new DB info.

* Story/aad 44 (#11)

* Creating Info page for each shelter

Creating the get more info page for each of the shelters that are listed on the results page. Further styling is needed, and a button to direct to the mapview is needed.

* Update results.blade.php

Slugging the urls on the result page to make url more readable

* Adding images for shelters

These are temporary images for each of the shelters. Will need to talk to sponsors to get better sourced images

* Fixing image loading

Wrote a new function for getting images loaded using a variable and with disregard for file extension. These images will need to live in a different location.

* Update WelcomeController.php

Fixed images that had parentheses in their name

* Update WelcomeController.php

Fixed apostrophes

* Update WelcomeController.php

Removed all non alphanumeric chars due to special opening and closing quote characters.

* Built out the Get More Info Blade

Created and began basic styling of the get more info page that is can be navigated to from the results page.

Co-authored-by: Bishop <59637502+Iwasson@users.noreply.github.com>

* Add map page (#12)

* Getting coordinate data out of the browser

Console logs the users current location. This will prompt the user to allow their location data being used

* Rebasing off of develop

Updating the branch to now have the changes from develop. Outputs current coords to the console for use in search filters

* Setting up mapbox

Getting basic mapbox displaying and setting up the files needed to run everything

* Adding navigation controls

Added navigation controls to the map, allowing users to input a starting and finishing location and choosing which mode of transportation they want to use. Need to investigate if there is a way to get trimet data for transit lines

* Generate map of shelter locations

* Add shelter marker popup features

* Make route directions responsive

* Integrate map with info page

Co-authored-by: Bishop <59637502+Iwasson@users.noreply.github.com>

* Aad 8 (#15)

* AAD-008: change search button style

* AAD-008: change search button style

* AAD-008:Add Carousel in homepage

* AAD-8: add map card and information card

* add top card

* Add description at main page

* ADD-8 add black borders to text

* AAD-8 add contect

* AAD-8 add gap

Co-authored-by: Iwasson <59637502+Iwasson@users.noreply.github.com>

* Added additional location options to get current locations (#13)

* Getting coordinate data out of the browser

Console logs the users current location. This will prompt the user to allow their location data being used

* Rebasing off of develop

Updating the branch to now have the changes from develop. Outputs current coords to the console for use in search filters

* Setting up mapbox

Getting basic mapbox displaying and setting up the files needed to run everything

* Adding navigation controls

Added navigation controls to the map, allowing users to input a starting and finishing location and choosing which mode of transportation they want to use. Need to investigate if there is a way to get trimet data for transit lines

* Generate map of shelter locations

* Add shelter marker popup features

* Make route directions responsive

* Integrate map with info page

* Added additional options to get current location

Co-authored-by: Bishop <59637502+Iwasson@users.noreply.github.com>

* Story/aad 18 (#14)

* Update welcome.blade.php

Fixed typo in welcome statement.

* Update results.blade.php

Removed address from results view page.

* Update SheetsController.php

Added 'Beds Available:' statement to print out for the number of beds. Removed address variable from the result array. These are held in the result array as variables for the controller to share with view pages.

* Update information.blade.php

Fixed typo from 'avalible' to 'available'.

* Update results.blade.php

Changed another typo for the results page from 'avalible' to 'available'.

* Update welcome.blade.php

Typo fix

* Fixed more typos and inspected css.

Fixed several more typos.

Fixed images for mobile scale.

Co-authored-by: Bishop <59637502+Iwasson@users.noreply.github.com>

* Consistency pass (#16)

* First pass on main page

* Completed consistency pass

Made each page utilize the projects color-scheme and styling.

* Further formatting and styling

* Fixed typo

Co-authored-by: zchronis <60366837+zchronis@users.noreply.github.com>

Co-authored-by: zchronis <60366837+zchronis@users.noreply.github.com>
Co-authored-by: Matthew Pickett <91637250+matt-pickett@users.noreply.github.com>
Co-authored-by: Yuxin <74965440+sunyux@users.noreply.github.com>
Co-authored-by: TimmyYoung <77657486+TimmyYoung@users.noreply.github.com>