Additional map styling (#20)
* Getting coordinate data out of the browser

Console logs the users current location. This will prompt the user to allow their location data being used

* Rebasing off of develop

Updating the branch to now have the changes from develop. Outputs current coords to the console for use in search filters

* Setting up mapbox

Getting basic mapbox displaying and setting up the files needed to run everything

* Adding navigation controls

Added navigation controls to the map, allowing users to input a starting and finishing location and choosing which mode of transportation they want to use. Need to investigate if there is a way to get trimet data for transit lines

* Generate map of shelter locations

* Add shelter marker popup features

* Make route directions responsive

* Integrate map with info page

* Added additional options to get current location

* Additional styling and coordinating map colors to website

Co-authored-by: Bishop <59637502+Iwasson@users.noreply.github.com>